[embind] Change GenericEnumValue's type to int (#20908)
`long` is interpreted as a pointer and causes negative `enum class`
values to be casted to unsigned and become large overflown values when
`CAN_ADDRESS_2GB` is true:
https://github.com/emscripten-core/emscripten/blob/0a3b7539856df97370c2e42e174850883c735c9a/src/jsifier.js#L153

+++ system/include/emscripten/bind.h
@@ -37,7 +37,7 @@ enum class sharing_policy {
 
 namespace internal {
 
-typedef long GenericEnumValue;
+typedef int GenericEnumValue;
 
 typedef void* GenericFunction;
 typedef void (*VoidFunctionPtr)(void);
